NEXT SUCCESS VERSION
This version has gotten successfully the energy deposition by  the photons, the global time of muons and local time of photons; also the wavelenght and the  number of photons produced. All the data could be viewed using ROOT. 
Also the geometries are in their correct position and the right dimensions.